refactor: add logic in delete drivers and automobiles

+++ src/application/useCases/automobiles/DeleteAutomobileUseCase.ts
@@ -1,8 +1,9 @@
-import { NotFoundError } from "../../../domain/error/AppError";
+import { BadRequestError, NotFoundError } from "../../../domain/error/AppError";
 import { IAutomobileRepository } from "../../../domain/repositories/AutomobileRepository";
 import { GetByIdAutomobileUseCase } from "./GetByIdAutomobileUseCase";
 import { deleteAutomobileSchema } from "./automobileSchemaInput";
 import { validator } from "../../../domain/validation";
+import { IAllocationRepository } from "../../../domain/repositories/allocationRepository";
 
 interface DTO {
 	id: string;
@@ -11,7 +12,8 @@ interface DTO {
 export class DeleteAutomobileUseCase {
 	constructor(
 		private automobileRepository: IAutomobileRepository,
-		private getByIdAutomobileUseCase: GetByIdAutomobileUseCase
+		private getByIdAutomobileUseCase: GetByIdAutomobileUseCase,
+		private allocationRepository: IAllocationRepository
 	) {}
 
 	async handle(args: DTO): Promise<void> {
@@ -19,6 +21,10 @@ export class DeleteAutomobileUseCase {
 		const automobile = await this.getByIdAutomobileUseCase.handle(validatedData);
 		if (!automobile) {
 			throw new NotFoundError('Automobile not found');
+		}
+		const isAlreadyAllocated = await this.allocationRepository.findByAllocationByAutomobileIdInProgress(validatedData.id);
+		if (isAlreadyAllocated) {
+			throw new BadRequestError('Automobile is already allocated to a driver, please finish the allocation before deleting the automobile');
 		}
 		await this.automobileRepository.delete(validatedData.id);
 	}
