+++ src/application/useCases/drivers/DeleteDriverUseCase.ts
@@ -1,8 +1,9 @@
-import { NotFoundError } from "../../../domain/error/AppError";
+import { BadRequestError, NotFoundError } from "../../../domain/error/AppError";
 import { IDriverRepository } from "../../../domain/repositories/DriverRepository";
 import { GetDriverByIdUseCase } from "./GetDriverByIdUseCase";
 import { deleteDriverSchema } from "./driverSchemaInput";
 import { validator } from "../../../domain/validation";
+import { IAllocationRepository } from "../../../domain/repositories/allocationRepository";
 
 interface DTO {
 	id: string;
@@ -11,7 +12,8 @@ interface DTO {
 export class DeleteDriverUseCase {
 	constructor(
 		private driverRepository: IDriverRepository,
-		private getDriverByIdUseCase: GetDriverByIdUseCase
+		private getDriverByIdUseCase: GetDriverByIdUseCase,
+		private allocationRepository: IAllocationRepository
 	) {}
 
 	async handle(args: DTO): Promise<void> {
@@ -19,6 +21,10 @@ export class DeleteDriverUseCase {
 		const driver = await this.getDriverByIdUseCase.handle(validatedData);
 		if (!driver) {
 			throw new NotFoundError('Driver not found');
+		}
+		const isAlreadyAllocated = await this.allocationRepository.findByAllocationByAutomobileIdInProgress(validatedData.id);
+		if (isAlreadyAllocated) {
+			throw new BadRequestError('Driver is already allocated to an automobile, please finish the allocation before deleting the driver');
 		}
 		await this.driverRepository.delete(validatedData.id);
 	}
